+++ src/main/java/com/acme/yupanaapi/domain/service/FlowService.java
@@ -12,12 +12,12 @@ import com.acme.yupanaapi.resource.UserWalletResource;
 import org.springframework.http.ResponseEntity;
 
 public interface FlowService {
-    Flow getFlowById(Long flowId);
-    Flow createFlow(Flow flow, Long walletId, Long sellerId);
-    Flow updateFlow(Long flowId,Long walletId, Long sellerId, Flow flowRequest);
-    ResponseEntity<?> deleteFlow(Long flowId);
-    List<Flow> getAllByWalletId(Long walletId);
-    List<Flow> getAllByWalletIdAndDeadlineDate(Long walletId, Date date);
-    Flow getLastFlow(Long walletId);
-    UserWalletResource getData(Long walletId);
+    Flow getFlowById(Integer flowId);
+    Flow createFlow(Flow flow, Integer walletId, Integer sellerId);
+    Flow updateFlow(Integer flowId,Integer walletId, Integer sellerId, Flow flowRequest);
+    ResponseEntity<?> deleteFlow(Integer flowId);
+    List<Flow> getAllByWalletId(Integer walletId);
+    List<Flow> getAllByWalletIdAndDeadlineDate(Integer walletId, Date date);
+    Flow getLastFlow(Integer walletId);
+    UserWalletResource getData(Integer walletId);
 }
